refactor: Extract authentication and canvas management logic into new `useAuth` and `useCanvas` hooks.
